Fix ARM32 format/type portability in cellframe and update dap-sdk ref

+++ modules/type/dag/dap_chain_cs_dag.c
@@ -690,7 +690,7 @@ static bool s_chain_callback_datums_pool_proc(dap_chain_t *a_chain, dap_chain_da
         HASH_CLEAR(hh_select, l_tmp);
         pthread_mutex_unlock(&PVT(l_dag)->events_mutex);
         if (l_hashes_linked < l_hashes_size) {
-            log_it(L_ERROR, "No enough unlinked events present (only %lu of %lu), a dummy round?", l_hashes_linked, l_hashes_size);
+            log_it(L_ERROR, "No enough unlinked events present (only %zu of %zu), a dummy round?", l_hashes_linked, l_hashes_size);
             return false;
         }
     }
@@ -700,7 +700,7 @@ static bool s_chain_callback_datums_pool_proc(dap_chain_t *a_chain, dap_chain_da
      * or we have successfully chosen the hash(es) to link with.
      * No additional conditions required.
     */
-    uint64_t l_event_size = 0;
+    size_t l_event_size = 0;
     dap_chain_cs_dag_event_t * l_event = l_dag->callback_cs_event_create
             ? l_dag->callback_cs_event_create(l_dag, a_datum, l_hashes, l_hashes_linked, &l_event_size)
             : NULL;
